% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adpss.R
\docType{package}
\name{adpss-package}
\alias{adpss-package}
\title{adpss: A package for design and analysis of a locally or globally efficient adaptive design involving adaptive sample size determination.}
\description{
This package provides the functions for conducting a trial with adaptive
sample size determination.
Such a trial consists of adaptive determination of sample size
at an interim analysis and implementation of frequentist statistical test of
a prefixed significance level.
The required assumptions for the stochastic process of the test statistics is Brownian motion.
Predetermination of adaptation rule is not required, i.e., adaptations can be made with full flexibility.
}
